fix(compiler): make assingment `lvalue` a reference instead of expression  (#3734)
This is a small cleanup. It reverts a change done in #3213 where we added phases to all expressions and in the process turned assignment targets into expressions. In practice it doesn't seem the assignment target itself needs a phase other than the phase already assigned to the inner expression in the reference and the phase assigned to the variable itself.
This also removed a duplicate error message when trying to access `this` from a static function due to double resolving caused by resolving the entire expression and then the inner expression.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.