fix(sdk): fixing `api.url` token inconsistency (#3772)
Fixes: #3024 #3539 #3342 
The problem with the token was the way cdktf is tokenizing strings, I assume it tried to create a new token every time we call `api.url`, or maybe tokenize a token. Now they're all the same- no matter if you call it on preflight or inflight it or assign it to another variable- it'll generate exactly the same token number.

other than that:
- brought back the aws api SDK tests
- changed sim response to be an empty string (in express it's an empty object by default)

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.