chore(sdk): refactor connections API (#3814)
Background: in order for Wing classes to support extending non-Wing classes (#3717), I'm trying to remove the `Resource` class in the SDK so that all classes generated by our compiler just need to extend the ol' [`construct.Construct`](https://github.com/aws/constructs/tree/10.x) class -- the "base class" of the JSII ecosystem. This is a tricky effort since `Resource` is used in many places, and it collected a lot of responsibilities over the past year.

This PR takes a first step towards that by removing the connection-related information from the `Resource` class:

https://github.com/winglang/wing/blob/1c9954e50ebdd4f2b290e91665c920fb504c9cb7/libs/wingsdk/src/std/resource.ts#L222-L223

Instead, we now store the information inside the root of the construct tree. It can be accessed through `Connections.of(x)` where `x` is any class instance. See `connections.ts` for the full implementation (its short).

With this refactoring, I took the opportunity to clean up how connections are modeled. Previously, we were storing each connection twice inside `tree.json` (once next to the connection's source, and a second time next to the connection's target). Now, we just store each connection once, in a flat array inside `connections.json`. (I wasn't sure if it makes more sense to put it in a separate file - I'm interested in feedback on this). I also removed the `implicit` field which wasn't used by the SDK or console anywhere.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.