chore(repo): ensure self mutation updates branch as well (#3922)
Pull requests do not actually run workflows on the head of a given branch, it uses the merge commit between that head and the target head (e.g. `main`)

This means that when mutation patches are generated, they assume they're operating on the merge commit. This PR makes that assumption true when applying patches by also updating the branch itself.

adding do-not-merge to test, but feel free to review.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.