chore(repo): ensure the self mutation updater actual does a real merge commit (#3946)
The change introduced in https://github.com/winglang/wing/pull/3922 technically works, but the result is confusing because the patch applied is not an actual merge commit. So Github shows those changes as if you made them.

This PR instead does the mutation update via the github API. The downside of this (other than the obvious complexity) is that technically the updated branch can be newer than the one the diffs were created against. I would rather have that problem than the one we have now.

Change has been tested in my fork here https://github.com/MarkMcCulloh/wing-distributed-workflow/pull/5

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.