chore(compiler): make preflight.js more self-contained (#3972)
Currently the `preflight.js` files generated by Wing contain a reference to a value named `$plugins` which has to be injected by a NodeJS VM. Because of this, it's not possible to run `node path/to/.wing/preflight.js` for debugging purposes without manually editing the file to stub in a fake value.

I've removed the need to inject this value by passing the list of plugin paths through an environment variable instead.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.