fix(compiler): missing extended struct fields for some imported JSII interfaces (#3982)
Fixes #3716

The core of this change is to make sure that when importing an interface, the type itself is registered before importing the types that it extends. Previously it was the opposite way. 
I will say that I'm not sure why this works, the previous implementation makes more sense to me 🤷 

No tests were added because I was not able to replicate the issue in other JSII libs other than `cdk8s-plus-27` so I wasn't sure that it's worthwhile to add an additional package just to test this. We also don't have a fixture system for JSII testing.

Misc:
 - Made sure the imported (actual) interfaces also get their extended properties. Not sure if this was intentionally left out before

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.