chore(repo): skip hangar when not needed (#3791)
### Goal
Skip e2e tests when the changes in the PR don't effect it. Notably, this usually means unit tests, README, or anything in the console.

"changes" here means one of 2 things:
- **If non-pr:** Last successful build workflow on the same branch
- **If pr:** git diff of the PR compared to the base, ignoring any commits from upstream

This high-level logic is the same that we've been using to determine whether or not we deploy the vscode extension, so this extends it to the e2e and benchmarking tests as well.
Unfortunately turbo itself was not sufficient for this goal (see note at the top of `tools/bump-pack/src/turbo-diff.ts`), so I had to write some junk to properly determine "what changed". I swear I've had to write something like this like 9 times already at various points in my life.

#### Limitations

The "Last successful build workflow on the same branch" does not apply to PRs, even though I wish it did. Implementing this safely is more difficult than it seems.

### The future of bump-pack

Part of the this PR is also effectively turning "bump-pack" into a general-purpose toolkit. In a followup PR I plan to simply turn the entire package into a full CLI to serve misc repo development purposes. My hope is that this will make it easier to actually add testing to it with a fixture based on https://github.com/kiegroup/mock-github and https://github.com/kiegroup/act-js

### Misc

- Added a script for the git patching stuff, didn't like so much repetitious bash in the workflow
- ~Effectively forked `nrwl/nx-set-shas` since it was not sufficient for the workflow I wanted here (It always assumed you wanted to compare against `main` in a PR)~ Nevermind, it turns out that the original action is actually good enough


*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.