fix(sdk)!: change Json.get() and Json.getAt() to throw (#4034)
Currently `Json.get()` and `Json.getAt()` don't throw any error if the JSON property you are accessing doesn't exist.

Besides, the descriptions of the methods say that they can return `undefined`, which is not true: it's not specified in the return values and that's actually the scope of `Json.tryGet()` and `Json.tryGetAt()`.

- [x] Fix `Json.get()`
- [x] Fix `Json.getAt()`

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.