feat(console):  improve jump to error when runing in vscode (#4012)
When opening a hyperlink in vs-code, the default behavior is to open the hyperlink in the currently focused 'Editor Group.' 

However, when we have the Console open side by side with the wing file, this default behavior causes vs-code to open the file as a new tab in the Console panel group (if the group where you have the Console is focused).

This PR adds two new endpoints: `[POST] app.openFileInEditor` and `app.openFileInEditorSubscription`
When clicking on an error link the Console emits an event named `openFileInEditor`.
Code editors like vscode (and potentially others) can subscribe to the new sub endpoint to receive these events.

Previous behaviour (using hyperlinks)

https://github.com/winglang/wing/assets/5547636/305a58ef-1f90-499e-a72d-60a00cefe2ca

New behaviour (using sub)

https://github.com/winglang/wing/assets/5547636/ae9c5d84-c70a-4160-b019-7b9144509797

Weird case when you have 425924593 groups

https://github.com/winglang/wing/assets/5547636/ec52da55-887d-4b46-84a2-8d0b566669b0

resolves https://github.com/winglang/wing/issues/3337

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.