feat(sdk): only generate test-running resources on clouds when running `wing test` (#4098)
It was brought up in a recent discussion about testing that it's unusual how `test` statements create resources in cloud provider accounts even when you have only compiled your application to run in production. This PR changes the behavior so that test resources are only generated when running `wing test`, or when using the simulator target (since there's no cost of simulated resources). 

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.