fix(build): missing files @winglang/sdk - cannot find module (#4088)
fixes  #4021

When `libs/wingsdk/src/.gen/providers` directory is not empty, it fails to generate the required files during `pre-compile` stage. The easy - but not obvious - fix is to remove *providers* directory entirely.

`pnpm pre-compile` command executed `cdktf get` command, which does not generate files if it spots the aforementioned directory.  This command has a `--force` flag, which forces to generate all the files. I changed `pre-compile` script to use it by default.

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [x] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.