fix: ensure wing process is killed from vscode (#4124)
Fixes #4119

Since wing itself is being run in a child process of vscode, it wasn't handling the kill signal. Detaching it allows it to have its own process tree and be killed more reliably. Special handling is needed for windows, of course.

This also ensures that any process spawned by `wing it` itself is terminated too.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.