feat(vscode): automatically reload extension features when updating wing (#4094)
Fixes #4006
Fixes #4022

This is best-effort approach, where we watch the specific winglang binary for changes (followed symlinks). It's difficult to check how this works on other systems, but it works on my machine (haha). After this ships I was hoping to try it on other machines. 
It also works when changing the configuration in the extension settings.

Misc:
- Fixed using the wing.bin configuration to set a custom wing binary
- Unified logging in the extension
- Added more logging, mostly for error cases
- Removed the ability to use `npx` for the wing binary. I double people are using it and it's not really a great way to use wing anyways

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.