fix(sdk): cloud.Queue is not FIFO by default (#4173)
Related to https://github.com/winglang/wing/issues/3897

We have a flakey test in our CI that validates part of the functionality of `cloud.Queue`. It turns out, our test was assuming that the queue messages are received in FIFO order (first-in-first-out), which wasn't something we meant to guarantee (see the original issue about `queue.pop()` [here](https://github.com/winglang/wing/issues/1385)). This PR fixes the test, and updates the simulator implementation to reflect the fact that messages can be received out of order.

(Why not FIFO by default? For many distributed apps, the exact order of messages isn't critical, and relaxing this constraint makes it easier to handle higher throughputs of messages. That said, adding a `fifo` option to `cloud.Queue` would be reasonable as a future enhancement)

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.