fix(sdk): cherrypick default az for wing app (#4151)
Currently, when a new Wing app is created for AWS targets, it's left up to AWS to choose a random AZ in which to deploy the private and public subnets. With this change, we tell AWS that we always want to deploy to the first AZ of whatever region specified by the user, which is generally the one that has the broadest range of services available.

Closes #4126

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.