fix(compiler): implicit structs don't work in collections (#4199)
Fixes #4185
Fixes #3231 (should be allowed to assign Map to Json, no need to change the error. Fixed specific bug which prevented this for `Json?`)

Most of the change here was in two areas:
- Made sure that collections of Json will always have the Json type data stored in them
- Extracted the json logic from validate_type_in to a new function because it needed some recursion. Heavily commented it now too. The overall logic is only slightly changed, though, to support nested collections.

You'll notice some duplication between the collection literals. If there is concern about it I'm cool with doing some consolidation in this PR, but I figured it could wait until https://github.com/winglang/wing/issues/3873

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.