fix(sdk): better error message on unsupported compile target (#4282)
closes https://github.com/winglang/wing/issues/4172

The new error message tells that the resource still needs to be implemented for the target.
It also includes a link to the roadmap with a filter, so user can find the ticket for the implementation

Before | After 
---|---
Unable to create an instance of abstract type "@winglang/sdk.cloud.Api" for this target | Resource "@winglang/sdk.cloud.Api" is not yet implemented for "awscdk" target. Please refer to the roadmap https://github.com/orgs/winglang/projects/3/views/1?filterQuery=cloud.Api

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.