feat: state in `cloud.Service`
Redesign the API for `cloud.Service` so it is possible to maintain state across `onStart` and `onStop`.

The API now accepts an object that implements `IServiceHandler` which declares two inflight methods: `onStart()` and `onStop()`.

The handler is emitted into a single JavaScript entrypoint which creates a single instance of the object and calls the callbacks on the same object.

This implies that inflight state is preserved throughout the lifetime of the service. This also means that an external JavaScript module can preserve state through a global variable.

Added a bunch of SDK tests which demonstrate this API.

Closes #2237
Closes #4248

Remaining Work:

- [ ] Update docs
- [ ] Migrate the rest of the SDK tests to Wing (it's becoming difficult to implement these handlers in JavaScript).