chore(compiler): update rust and wasi toolchain (#4458)
This was a difficult update. A newer version of rust exposed/caused an issue with our UnsafeRefs in some cases, causing memory corruption. I tried to highlight semi-functional changes

~Looks like the wasm size has increased by almost 2x. Not ideal, but not sure why or what to do. Based on benchmarks it looks like the perf is unchanged at least.~ Ok it turns out the new version of rust also somehow broke cargo-wasi and it no longer ran wasm-opt anymore. So now this PR also removes cargo-wasi and instead runs wasm-opt directly from the binaryen toolchain.
On the bright side, the install for this is much faster. The only other useful special thing that cargo-wasi did was run https://github.com/rustwasm/walrus to demangle names. Since we always do release builds, I'm not sure how valuable that is.

Updates:
- wasi-sdk: 19 -> 20
- rust: 1.67.1 -> 1.73.0
- cargo-wasi: Removed. Now using the `wasm32-wasi` target and wasm-opt directly
- tree-sitter: 0.20.7 -> 0.20.8
- All crates were updated
- All crates were set to use similar pinning as `^`, where patches are automatically installed

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.