feat(vscode): go to definition (#4522)
https://github.com/winglang/wing/assets/1237390/7a9449d5-f147-408c-b3ed-cbb82863726b

Fixes #2225

As part of adding this feature, I changed the HoverVisitor into the SymbolLocator. This also now uses VisitorContext to be more in line with other parts of the compiler. hover and goto_definition are both able to make use of this tool.

If you're wondering about the difference between "go to definition" and "go to declaration": For now they are lumped in together for simplicity, but there is value in separation. I'm willing to bet that this feature will satisfy what a lot of people are looking for right now.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.