fix(vscode): no symbol details for `extends` or `impl` (#4527)
Fixes #4526

Tried to minimize the work done in SymbolLocator so that the base visitor can make sure to handle cases like this. It adds a small bit of redundant traversal at runtime but makes it harder to hit edge cases if SymbolLocator gets out of sync with the base visitor implementation.

Other Fixes:
- hover for interface and struct declaration fields. It's not super useful, but it also felt weird that it didn't work

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.