fix(vscode): panic when hovering over std types (#4533)
A std type like `Json` is actually `std.Json` in the AST. That `std` part is not technically part of the file so it does not have an actual span. This panicked when comparing file_id in the merge function.

This change allows the span merge function to basically ignore default spans.

Sidenote: I'd really like to make WingSpan into an enum, but that change sounds like a lot of work to fix this panic.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.