feat(sdk)!: expose simulator via HTTP server (#4105)
This PR evolves the Wing simulator so that every simulation spins up an HTTP server. This HTTP server accepts requests to call inflight methods on resources, making it possible for simulated resources to interact across multiple processes on the same machine (such as from containers running on your system).

Since it's not possible to serialize all forms of data as HTTP, there's a small regression where the `Redis` class's `rawClient` API no longer worked. For the time being, I've removed the method to work around the issue. We might be able to recover the functionality by writing a dedicated implementation of `toInflight` for the `sim` target's implementation of `Redis`, but it's a bit beyond the scope of this PR.

BREAKING CHANGE: The `rawClient()` method on `ex.Redis` has been removed. This API was infrequently used and was provided without any type information. If you have a use case for this API, let us know!

## Checklist

- [x] Title matches [Winglang's style guide](https://www.winglang.io/contributing/start-here/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Wing Cloud Contribution License](https://github.com/winglang/wing/blob/main/CONTRIBUTION_LICENSE.md)*.